Cherrypick fixes from dev to pointrelease 25101 (#19392)
* Fix android license for AR (#19178)

* Add auto-accept step

Signed-off-by: Mike Chang <changml@amazon.com>

* Fix unused variable warning in Clang release build (#19299)

Signed-off-by: Markus Prettner <markus.prettner@huawei.com>

* Fix registration of wrong type for AZ::s64 in Qt (#19300)

Signed-off-by: Markus Prettner <markus.prettner@huawei.com>

* No longer require `git` to build and run in installer (#19355)

The v-hacd library had 2 different files that were
intended to be used, one for the code version of the
engine which downloaded v-hacd from github, using git,
and one for the installer version of the engine
which noted the copyright information and the fact
that it was being used, but didn't need to actually
download it since its entire contents were subsumed
inside the physx gem dlls and were not public api.

Its also only used by the tools (editor, etc) and
not the standalone game runtime, so it isn't necessary
to have it available in the installer build.

The original problem is caused by the use of the
cmake `ly_add_external_target_path` function, which
appears to make sure that the files in that location
are copied to the installer.

Instead, in this case, we want different files in the
installer than in the code version, so instead of using
that macro, we add the directory to the module path
directly, and then use `ly_install` to copy the other
cmake file (that does not use git) to the installer.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Made the ScriptCanvasApplication toggleable with a cmake setting (#19356)

* Made the ScriptCanvasApplication toggleable with a cmake setting
* PR feedback
---------

Signed-off-by: Luis Sempé <58790905+lsemp3d@users.noreply.github.com>

---------

Signed-off-by: Mike Chang <changml@amazon.com>
Signed-off-by: Markus Prettner <markus.prettner@huawei.com>
Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>
Signed-off-by: Luis Sempé <58790905+lsemp3d@users.noreply.github.com>
Co-authored-by: Mike Chang <changml@amazon.com>
Co-authored-by: Markus Prettner <markus.prettner@huawei.com>
Co-authored-by: Luis Sempé <58790905+lsemp3d@users.noreply.github.com>