Allows you to save default paths into assetbundler (#19419)
the previous version of assetbundler does not actually save
paths you specify in the GUI, since the intention there was
that the gui is for constructing rule files that have blanks
for you to fill in later, from the command line.

This change makes it so that the gui actually saves those paths
into the asset files, if you want.  This would allow a smaller
(1 user) team to use the gui to assemble bundles, while still allowing
power users to use the command line with custom paths to substitute.

You don't have to use this feature - you can just leave those fields
blank as usual.  But if you do fill them in, they will be saved
into the rules file, which allows you to run assetbundlerbatch and
fill in the `default` token (`$`) for those paths to just use the
ones from the file.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>